% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertions}
\alias{assertions}
\alias{assert_has_names}
\alias{assert_is_scalar_character}
\alias{assert_is_scalar_logical}
\alias{assert_is_existing_file}
\alias{assert_is_existing_dir}
\alias{assert_unique_id}
\title{Assertion Functions}
\usage{
assert_has_names(x, names, .arg = "x")

assert_is_scalar_character(x, .arg = "x")

assert_is_scalar_logical(x, .arg = "x")

assert_is_existing_file(path, .arg = "path")

assert_is_existing_dir(path, .arg = "path")

assert_unique_id(df, id_col)
}
\arguments{
\item{x}{Object to validate.}

\item{names}{Character vector of required names.}

\item{.arg}{Name of the argument being validated (for error messages).}

\item{path}{File or directory path.}

\item{df}{Data frame to check.}

\item{id_col}{Column name(s) that should be unique.}
}
\value{
Invisibly returns \code{TRUE} if assertion passes; otherwise errors.
}
\description{
Fast-fail validation helpers with actionable error messages.
}
\examples{
\dontrun{
assert_has_names(list(a = 1, b = 2), c("a", "b"))
assert_is_scalar_character("test")
assert_is_scalar_logical(TRUE)
}

}
