% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{new_niche_recipe}
\alias{new_niche_recipe}
\title{Create a niche_recipe object}
\usage{
new_niche_recipe(x = list())
}
\arguments{
\item{x}{List containing recipe data}
}
\description{
Create a niche_recipe object
}
