% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{new_niche_spec}
\alias{new_niche_spec}
\title{Create a niche_spec object}
\usage{
new_niche_spec(x = list())
}
\arguments{
\item{x}{List containing specification data}
}
\description{
Create a niche_spec object
}
