% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{niche_abort}
\alias{niche_abort}
\title{Abort with niche-style error}
\usage{
niche_abort(message, details = NULL, help = NULL)
}
\arguments{
\item{message}{Error message}

\item{details}{Optional list of details}

\item{help}{Optional help text}
}
\description{
Abort with niche-style error
}
