% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{niche_output_paths}
\alias{niche_output_paths}
\title{Create output directory paths}
\usage{
niche_output_paths(root = "outputs")
}
\arguments{
\item{root}{Root output directory}
}
\description{
Create output directory paths
}
