% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{validate_niche_recipe}
\alias{validate_niche_recipe}
\title{Validate a niche_recipe object}
\usage{
validate_niche_recipe(recipe)
}
\arguments{
\item{recipe}{A niche_recipe object}
}
\description{
Validate a niche_recipe object
}
