% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{write_audit_json}
\alias{write_audit_json}
\title{Write canonical JSON for audit trail}
\usage{
write_audit_json(x, path)
}
\arguments{
\item{x}{Object to write}

\item{path}{Output path}
}
\description{
Write canonical JSON for audit trail
}
