% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{assertions}
\alias{assertions}
\alias{assert_has_names}
\alias{assert_is_scalar_character}
\alias{assert_is_scalar_logical}
\alias{assert_is_existing_file}
\title{Assertion Functions}
\usage{
assert_has_names(x, names, .arg = "x")

assert_is_scalar_character(x, .arg = "x")

assert_is_scalar_logical(x, .arg = "x")

assert_is_existing_file(path, .arg = "path")
}
\arguments{
\item{x}{Object to validate.}

\item{names}{Character vector of required names.}

\item{.arg}{Name of the argument being validated (for error messages).}

\item{path}{File path.}
}
\value{
Invisibly returns \code{TRUE} if assertion passes; otherwise errors.
}
\description{
Fast-fail validation helpers with actionable error messages.
}
\examples{
\dontrun{
assert_has_names(list(a = 1, b = 2), c("a", "b"))
assert_is_scalar_character("test")
assert_is_scalar_logical(TRUE)
}

}
